requirements.txt
streamlit>=1.30.0
PyPDF2>=3.0.0
python-docx>=1.1.0
plotly>=5.18.0
